+++ src/handleUpload.ts
@@ -9,6 +9,7 @@ import path from 'path'
 interface Args {
   acl?: 'private' | 'public-read'
   bucket: string
+  cacheControl?: string
   collection: CollectionConfig
   getStorageClient: () => AWS.S3
   prefix?: string
@@ -20,6 +21,7 @@ const multipartThreshold = 1024 * 1024 * 50
 export const getHandleUpload = ({
   acl,
   bucket,
+  cacheControl,
   getStorageClient,
   prefix = '',
 }: Args): HandleUpload => {
@@ -36,6 +38,7 @@ export const getHandleUpload = ({
         ACL: acl,
         Body: fileBufferOrStream,
         Bucket: bucket,
+        CacheControl: cacheControl,
         ContentType: file.mimeType,
         Key: fileKey,
       })
@@ -50,6 +53,7 @@ export const getHandleUpload = ({
         ACL: acl,
         Body: fileBufferOrStream,
         Bucket: bucket,
+        CacheControl: cacheControl,
         ContentType: file.mimeType,
         Key: fileKey,
       },